chore: reorganised utils
