+++ src/hooks.server.js
@@ -6,7 +6,7 @@
 */
 import { sequence } from "@sveltejs/kit/hooks"
 
-import { settings }  from "$lib/settings"
+import { settings }  from "$lib/settings.js"
 // #endregion
 
 
@@ -15,9 +15,9 @@ import { settings }  from "$lib/settings"
     // #region auth()
         // #region Specific Imports
 import dbClient from "$lib/server/database/prisma/prisma.js"
-import inputHandler from "$lib/server/inputHandler.js"
-import logError from "$lib/server/errorLogger.js"
-import { dateFromNow } from "$lib/utils/dateUtils"
+import inputHandler from "$lib/server/utils/inputHandler.js"
+import logError from "$lib/server/utils/errorLogger.js"
+import { dateFromNow } from "$lib/client/utils/dateUtils.js"
         // #endregion
