+++ src/lib/server/database/prisma/seed.js
@@ -1,6 +1,13 @@
 // #region Imports
 import dbClient from "./prisma.js"
-import inputHandler from "../../inputHandler.js"
+import inputHandler from "../../utils/inputHandler.js"
+/*
+    ISSUE: "$lib" alias does not work.
+
+    FIX: 
+        Create a script to run this file with vite,
+        since vite resolves the aliases.
+*/
 // #endregion
