+++ src/lib/server/utils/errorLogger.js
@@ -1,6 +1,6 @@
 // #region Imports
 import dbClient from "$lib/server/database/prisma/prisma.js"
-import dataUtils from "../utils/dateUtils.js"
+import dataUtils from "$lib/client/utils/dateUtils.js"
 import chalk from "chalk"
 // #endregion
